Fix nas rotas de geolocalizacao

+++ src/domain/portal/application/use-cases/geolocation/delete-geolocation-use-case.ts
@@ -1,14 +1,20 @@
 import { Either, left } from "src/core/Either"
 import { GeolocationNotFound } from "../../errors/geolocation/geolocation-not-found"
+import { ProfileActionNotAllowed } from "../../errors/profile/ProfileActionNotAllowed"
 
-type DeleteGeolocationUseCaseUseCaseResponse = Either<GeolocationNotFound, void>
+type DeleteGeolocationUseCaseUseCaseResponse = Either<
+  GeolocationNotFound | ProfileActionNotAllowed,
+  void
+>
 
 export class DeleteGeolocationUseCase {
   constructor(private mapRadiusRepository) {}
 
   async execute(
     mapRadiusId: number,
+    isCompany: boolean,
   ): Promise<DeleteGeolocationUseCaseUseCaseResponse> {
+    if (isCompany) return left(new ProfileActionNotAllowed())
     const mapRadius = await this.mapRadiusRepository.fetchById(mapRadiusId)
     if (!mapRadius) return left(new GeolocationNotFound())
     await this.mapRadiusRepository.delete(mapRadiusId)
